significant changes to code, new private method handles cpu betting, emplty seats are now filled with type `None` instead of empty player objects. MOved some helpers to a new file titled 'misc_functions'.
Dealer is no longer a player at the table, but is now a part of the table class. Dealer is now a subclass of the Player class, which also now has the deck, rather than the deck being a separate entity. All deck functionality are now methods in the Dealer class. This means the dealer turn is now outside the main gameplay loop.

Finally, minor changes in how data is printed to console to prevent duplicates and reduce the number of print statements, also removed unused `getSeatPositions` function.

Seat select now  displays empty seats instead of occpied ones (bug).

`checkPlayerBust` function is now replaced with `isOver` method, `shouldHit` function now handles the logic for cpus.